soggybud water model, greenhouse scan safety and safari zone improvements

+++ src/main/kotlin/org/magic/magicaddons/data/greenhouse/Codecs.kt
@@ -101,7 +101,7 @@ object Codecs {
                 GREENHOUSE_SLOT_CODEC.optionalFieldOf("slot")
                     .forGetter { Optional.ofNullable(it.slot) },
 
-                Codec.INT.optionalFieldOf("waterLevel")
+                Codec.DOUBLE.optionalFieldOf("waterLevel")
                     .forGetter { Optional.ofNullable(it.waterLevel) },
 
                 GROWTH_STAGE_INFO_CODEC.optionalFieldOf("growthStage")
